# docs/technical/openapi.yaml
openapi: 3.0.3
info:
  title: DMLog API
  description: |
    A comprehensive Dungeons & Dragons campaign management system with AI-powered character automation.

    ## Overview
    DMLog provides a complete solution for Dungeon Masters and players to manage campaigns, characters, sessions, and more. Built with modern async Python patterns, it offers real-time features through WebSockets and comprehensive monitoring capabilities.

    ## Key Features
    - **Character Management**: Create, manage, and evolve D&D characters with full 5e support
    - **Campaign Organization**: Track multiple campaigns with session scheduling and world-building tools
    - **Real-Time Sessions**: WebSocket-based live gaming with dice rolling, combat tracking, and more
    - **AI Integration**: Powered character automation and decision making (planned)
    - **Analytics & Monitoring**: Comprehensive metrics and performance tracking

    ## Authentication
    Currently uses simple API key authentication. Production deployment will implement OAuth 2.0.

    ## Rate Limiting
    API endpoints are rate-limited to 60 requests per minute per client.

    ## WebSocket Support
    Real-time features are available through WebSocket connections at `/ws/connect`.

    ## Versioning
    Current version: v1. All endpoints are prefixed with `/api/v1/`.

    ## Support
    - Documentation: https://docs.dmlog.com
    - GitHub Issues: https://github.com/dmlog/dmlog/issues
...
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/v1/characters/:
    get:
      tags:
        - Characters
      summary: List characters
      description: Retrieve a paginated list of characters with optional filtering
      operationId: list_characters
      parameters:
        - name: skip
          in: query
          description: Number of characters to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of characters to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: search
          in: query
          description: Search term for character names
          required: false
          schema:
            type: string
        - name: race
          in: query
          description: Filter by character race
          required: false
          schema:
            type: string
            enum: [Human, Elf, Dwarf, Halfling, Gnome, Dragonborn, Half-Orc, Tiefling]
        - name: character_class
          in: query
          description: Filter by character class
          required: false
          schema:
            type: string
            enum: [Fighter, Wizard, Cleric, Rogue, Ranger, Paladin, Bard, Druid, Monk, Sorcerer, Warlock, Barbarian]
        - name: is_active
          in: query
          description: Filter by active status
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of characters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Characters
      summary: Create character
      description: Create a new D&D character with full stats and background
      operationId: create_character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterCreate'
      responses:
        '201':
          description: Character created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Character name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Character name already exists"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/campaigns/:
    get:
      tags:
        - Campaigns
      summary: List campaigns
      description: Retrieve a list of all campaigns
      operationId: list_campaigns
      parameters:
        - name: skip
          in: query
          description: Number of campaigns to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of campaigns to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: is_active
          in: query
          description: Filter by active status
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Campaigns
      summary: Create campaign
      description: Create a new campaign
      operationId: create_campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignCreate'
      responses:
        '201':
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    CharacterCreate:
      type: object
      required:
        - name
        - race
        - character_class
        - level
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Character name
          example: "Aragorn"
        race:
          type: string
          enum: [Human, Elf, Dwarf, Halfling, Gnome, Dragonborn, Half-Orc, Tiefling]
          description: Character race
          example: "Human"
        character_class:
          type: string
          enum: [Fighter, Wizard, Cleric, Rogue, Ranger, Paladin, Bard, Druid, Monk, Sorcerer, Warlock, Barbarian]
          description: Character class
          example: "Ranger"
        level:
          type: integer
          minimum: 1
          maximum: 20
          description: Character level
          example: 5
        background:
          type: string
          maxLength: 1000
          description: Character background story
          example: "Raised in the wilds, skilled tracker and hunter"
        alignment:
          type: string
          description: Character alignment
          example: "Chaotic Good"
        ability_scores:
          $ref: '#/components/schemas/AbilityScores'
        hit_points:
          type: integer
          minimum: 1
          description: Current hit points
          example: 42
        max_hit_points:
          type: integer
          minimum: 1
          description: Maximum hit points
          example: 45
        armor_class:
          type: integer
          minimum: 10
          description: Armor class
          example: 16
        speed:
          type: integer
          minimum: 1
          description: Movement speed in feet
          example: 30
        proficiency_bonus:
          type: integer
          minimum: 0
          description: Proficiency bonus
          example: 3
        skills:
          type: object
          description: Character skills and proficiencies
          example:
            stealth: 7
            perception: 5
            survival: 6
        equipment:
          type: array
          items:
            type: string
          description: Character equipment
          example: ["Longbow", "Leather armor", "Dagger"]
        spells:
          type: array
          items:
            type: string
          description: Known spells
          example: ["Hunter's Mark", "Cure Wounds"]
        is_active:
          type: boolean
          default: true
          description: Whether the character is currently active
    AbilityScores:
      type: object
      required:
        - strength
        - dexterity
        - constitution
        - intelligence
        - wisdom
        - charisma
      properties:
        strength:
          type: integer
          minimum: 1
          maximum: 20
          description: Strength score
          example: 16
        dexterity:
          type: integer
          minimum: 1
          maximum: 20
          description: Dexterity score
          example: 14
        constitution:
          type: integer
          minimum: 1
          maximum: 20
          description: Constitution score
          example: 15
        intelligence:
          type: integer
          minimum: 1
          maximum: 20
          description: Intelligence score
          example: 12
        wisdom:
          type: integer
          minimum: 1
          maximum: 20
          description: Wisdom score
          example: 14
        charisma:
          type: integer
          minimum: 1
          maximum: 20
          description: Charisma score
          example: 13
...
    CharacterResponse:
      type: object
      required:
        - id
        - name
        - race
        - character_class
        - level
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Character UUID
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Character name
          example: "Aragorn"
        race:
          type: string
          description: Character race
          example: "Human"
        character_class:
          type: string
          description: Character class
          example: "Ranger"
        level:
          type: integer
          description: Character level
          example: 5
        background:
          type: string
          description: Character background story
          example: "Raised in the wilds, skilled tracker and hunter"
        alignment:
          type: string
          description: Character alignment
          example: "Chaotic Good"
        ability_scores:
          $ref: '#/components/schemas/AbilityScores'
        hit_points:
          type: integer
          description: Current hit points
          example: 42
        max_hit_points:
          type: integer
          description: Maximum hit points
          example: 45
        armor_class:
          type: integer
          description: Armor class
          example: 16
        speed:
          type: integer
          description: Movement speed in feet
          example: 30
        proficiency_bonus:
          type: integer
          description: Proficiency bonus
          example: 3
        skills:
          type: object
          description: Character skills and proficiencies
          example:
            stealth: 7
            perception: 5
            survival: 6
        equipment:
          type: array
          items:
            type: string
          description: Character equipment
          example: ["Longbow", "Leather armor", "Dagger"]
        spells:
          type: array
          items:
            type: string
          description: Known spells
          example: ["Hunter's Mark", "Cure Wounds"]
        is_active:
          type: boolean
          description: Whether the character is currently active
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    CharacterListResponse:
      type: object
      required:
        - characters
        - total
        - skip
        - limit
      properties:
        characters:
          type: array
          items:
            $ref: '#/components/schemas/CharacterResponse'
          description: List of characters
        total:
          type: integer
          minimum: 0
          description: Total number of characters matching the criteria
        skip:
          type: integer
          minimum: 0
          description: Number of characters skipped
        limit:
          type: integer
          minimum: 1
          description: Maximum number of characters returned
...
    CampaignCreate:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Campaign name
          example: "The Lost Mines of Phandelver"
        description:
          type: string
          maxLength: 2000
          description: Campaign description
          example: "A classic D&D adventure for beginner players"
        setting:
          type: string
          maxLength: 500
          description: Campaign setting
          example: "Forgotten Realms"
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
          default: intermediate
          description: Campaign difficulty
        max_players:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          description: Maximum number of players
        start_level:
          type: integer
          minimum: 1
          maximum: 20
          default: 1
          description: Starting character level
        end_level:
          type: integer
          minimum: 1
          maximum: 20
          description: Target ending level
        is_active:
          type: boolean
          default: true
          description: Whether the campaign is currently active
...
    CampaignResponse:
      type: object
      required:
        - id
        - name
        - description
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Campaign UUID
        name:
          type: string
          description: Campaign name
          example: "The Lost Mines of Phandelver"
        description:
          type: string
          description: Campaign description
          example: "A classic D&D adventure for beginner players"
        setting:
          type: string
          description: Campaign setting
          example: "Forgotten Realms"
        difficulty:
          type: string
          description: Campaign difficulty
          example: "beginner"
        max_players:
          type: integer
          description: Maximum number of players
          example: 5
        start_level:
          type: integer
          description: Starting character level
          example: 1
        end_level:
          type: integer
          description: Target ending level
          example: 5
        is_active:
          type: boolean
          description: Whether the campaign is currently active
        character_count:
          type: integer
          description: Number of characters in the campaign
          example: 4
        session_count:
          type: integer
          description: Number of sessions completed
          example: 8
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error description
          example: "Character not found"
        error_code:
          type: string
          description: Application-specific error code
          example: "CHARACTER_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        request_id:
          type: string
          description: Request identifier for debugging
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid request parameters"
            error_code: "INVALID_PARAMETERS"
            timestamp: "2024-01-22T10:30:00Z"
            request_id: "req_123456789"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Internal server error"
            error_code: "INTERNAL_ERROR"
            timestamp: "2024-01-22T10:30:00Z"
            request_id: "req_123456789"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []